+++ src/interrupts/keyboard.rs
@@ -1,218 +1,246 @@
 use crate::{
     ports::{self, Port},
+    startup,
     vga::{self, CursorShift},
 };
 use core::{
-    mem,
-    sync::atomic::{AtomicBool, Ordering},
+    fmt::Display,
+    sync::atomic::{AtomicU8, Ordering},
+};
+use pc_keyboard::{
+    DecodedKey, HandleControl, KeyCode, KeyState, Keyboard, Modifiers, ScancodeSet2,
+    layouts::Us104Key,
+};
+use ps2::{
+    Controller,
+    error::KeyboardError,
+    flags::{ControllerConfigFlags, KeyboardLedFlags},
 };
 
-/// The current state of the keyboard.
-static KEYBOARD: KeyboardState = KeyboardState::new();
+/// Circular scancode buffer. Each AtomicU8 represents a scancode.
+/// The genius idea of this buffer was taken from
+/// [`this video`](https://www.youtube.com/watch?v=dL0GO9SeBh0&list=PLUZozxlhse-NUto5JeJ0EDXEUFloWBdA).
+static KBD_BUF: [AtomicU8; 256] = [const { AtomicU8::new(0) }; 256];
 
-#[derive(Clone, PartialEq)]
-#[repr(u8)]
-#[allow(unused)]
-pub enum Key {
-    Escape = 0x01,
-    One,
-    Two,
-    Three,
-    Four,
-    Five,
-    Six,
-    Seven,
-    Eight,
-    Nine,
-    Zero,
+/// Index into the last handled scancode in the keyboard buffer.
+static KBD_RPTR: AtomicU8 = AtomicU8::new(0);
 
-    Minus,
-    Equals,
+/// Index into the last added scancode to the keyboard buffer.
+static KBD_WPTR: AtomicU8 = AtomicU8::new(0);
 
-    Q = 0x10,
-    W,
-    E,
-    R,
-    T,
-    Y,
-    U,
-    I,
-    O,
-    P,
+/// Whether shift is being held or not. Used as pc-keyboard's shift check is dodgy.
+///
+/// - Bit 0 - Left shift
+/// - Bit 1 - Right shift
+static SHIFT: AtomicU8 = AtomicU8::new(0);
 
-    BracketLeft,
-    BracketRight,
-    Enter = 0x1C,
+/// Enables external interrupts, disables mouse, runs some tests, 
+/// sets config, then sets the scancode and numlock LEDs.
+pub fn init() -> Result<(), KbdInitError> {
+    super::sti();
 
-    A = 0x1E,
-    S,
-    D,
-    F,
-    G,
-    H,
-    J,
-    K,
-    L,
+    // Safety: kbd_handler returns immediately if !SYS_INIT, so this is the only use of ports 0x60 & 0x64.
+    let mut controller = unsafe { Controller::new() };
 
-    SemiColon,
-    Quote,
-    Home = 0x29,
-    Backslash = 0x2B,
+    // Disable devices
+    KbdInitError::map("Disable keyboard", controller.disable_keyboard())?;
+    KbdInitError::map("Disable mouse", controller.disable_mouse())?;
 
-    Z,
-    X,
-    C,
-    V,
-    B,
-    N,
-    M,
+    // It doesn't matter if it's an err since we're just flushing the buffer
+    _ = controller.read_data();
 
-    Comma,
-    Fullstop,
-    ForwardSlash = 0x35,
+    // Tests
+    KbdInitError::map("Controller test", controller.test_controller())?;
+    KbdInitError::map("Keyboard test", controller.test_keyboard())?;
 
-    Space = 0x39,
-    F1 = 0x3B,
-    F2,
-    F3,
-    F4,
-    F5,
-    F6,
-    F7,
-    F8,
-    F9,
-    F10,
-    NumLock,
-    ScrollLock = 0x46,
-    F11 = 0xD7,
-    F12 = 0xD8,
+    // Config
+    let mut cfg = ControllerConfigFlags::all();
+    cfg.set(ControllerConfigFlags::DISABLE_KEYBOARD, false); // enable kbd
+    cfg.set(ControllerConfigFlags::ENABLE_MOUSE_INTERRUPT, false);
+    cfg.set(ControllerConfigFlags::ENABLE_TRANSLATE, false); // so scancode set 2 is actually scancode set 2
+    KbdInitError::map("Set config", controller.write_config(cfg))?;
 
-    Unknown,
+    // Re-enable keyboard
+    let mut kbd = controller.keyboard();
+    KbdInitError::map("Keyboard echo", kbd.echo())?;
+
+    // Scancode set 2 & Num Lock LEDs
+    KbdInitError::map("Set scancode", kbd.set_scancode_set(2))?;
+    KbdInitError::map("Set LEDS", kbd.set_leds(KeyboardLedFlags::NUM_LOCK))?;
+    KbdInitError::map("Reset keyboard", kbd.reset_and_self_test())?;
+
+    Ok(())
 }
 
-/// The action keys held or pressed on the keyboard.
-#[derive(Default)]
-struct KeyboardState {
-    ctrl: AtomicBool,
-    shift: AtomicBool,
-    alt: AtomicBool,
-    caps: AtomicBool,
+/// Error returned from `init`.
+pub struct KbdInitError {
+    cmd: &'static str,
+    err: KeyboardError,
 }
 
-impl KeyboardState {
-    /// Creates a new `KeyboardState`.
-    const fn new() -> Self {
-        KeyboardState {
-            ctrl: AtomicBool::new(false),
-            shift: AtomicBool::new(false),
-            alt: AtomicBool::new(false),
-            caps: AtomicBool::new(false),
-        }
-    }
-
-    /// Toggles the `caps` bool.
-    fn flip_caps(&self) {
-        self.caps.fetch_xor(true, Ordering::Relaxed);
-    }
-}
-
-impl TryFrom<Key> for char {
-    type Error = ();
-    fn try_from(key: Key) -> Result<Self, Self::Error> {
-        let shift = KEYBOARD.shift.load(Ordering::Relaxed);
-        let caps = KEYBOARD.caps.load(Ordering::Relaxed);
-        return match key {
-            Key::One => default_or_shift('1', '!', shift),
-            Key::Two => default_or_shift('2', '@', shift),
-            Key::Three => default_or_shift('3', '#', shift),
-            Key::Four => default_or_shift('4', '$', shift),
-            Key::Five => default_or_shift('5', '%', shift),
-            Key::Six => default_or_shift('6', '^', shift),
-            Key::Seven => default_or_shift('7', '&', shift),
-            Key::Eight => default_or_shift('8', '*', shift),
-            Key::Nine => default_or_shift('9', '(', shift),
-            Key::Zero => default_or_shift('0', ')', shift),
-
-            Key::Minus => default_or_shift('-', '_', shift),
-            Key::Equals => default_or_shift('=', '+', shift),
-            Key::BracketLeft => default_or_shift('[', '{', shift),
-            Key::BracketRight => default_or_shift(']', '}', shift),
-            Key::Enter => Ok('\n'),
-            Key::Space => Ok(' '),
-            Key::Comma => default_or_shift(',', '<', shift),
-            Key::Fullstop => default_or_shift('.', '>', shift),
-            Key::Home => default_or_shift('`', '~', shift),
-            Key::SemiColon => default_or_shift(';', ':', shift),
-            Key::Quote => default_or_shift('\'', '\"', shift),
-            Key::Backslash => default_or_shift('\\', '|', shift),
-            Key::ForwardSlash => default_or_shift('/', '?', shift),
-
-            Key::Q => default_or_shift('q', 'Q', shift || caps),
-            Key::W => default_or_shift('w', 'W', shift || caps),
-            Key::E => default_or_shift('e', 'E', shift || caps),
-            Key::R => default_or_shift('r', 'R', shift || caps),
-            Key::T => default_or_shift('t', 'T', shift || caps),
-            Key::Y => default_or_shift('y', 'Y', shift || caps),
-            Key::U => default_or_shift('u', 'U', shift || caps),
-            Key::I => default_or_shift('i', 'I', shift || caps),
-            Key::O => default_or_shift('o', 'O', shift || caps),
-            Key::P => default_or_shift('p', 'P', shift || caps),
-            Key::A => default_or_shift('a', 'A', shift || caps),
-            Key::S => default_or_shift('s', 'S', shift || caps),
-            Key::D => default_or_shift('d', 'D', shift || caps),
-            Key::F => default_or_shift('f', 'F', shift || caps),
-            Key::G => default_or_shift('g', 'G', shift || caps),
-            Key::H => default_or_shift('h', 'H', shift || caps),
-            Key::J => default_or_shift('j', 'J', shift || caps),
-            Key::K => default_or_shift('k', 'K', shift || caps),
-            Key::L => default_or_shift('l', 'L', shift || caps),
-            Key::Z => default_or_shift('z', 'Z', shift || caps),
-            Key::X => default_or_shift('x', 'X', shift || caps),
-            Key::C => default_or_shift('c', 'C', shift || caps),
-            Key::V => default_or_shift('v', 'V', shift || caps),
-            Key::B => default_or_shift('b', 'B', shift || caps),
-            Key::N => default_or_shift('n', 'N', shift || caps),
-            Key::M => default_or_shift('m', 'M', shift || caps),
-
-            _ => Err(()),
-        };
-
-        fn default_or_shift(default: char, shift: char, is_shift: bool) -> Result<char, ()> {
-            Ok(if is_shift { shift } else { default })
+impl KbdInitError {
+    /// Maps a `Result<T, E>` to a `Result<(), Self>`
+    fn map<T, E>(cmd: &'static str, err: Result<T, E>) -> Result<(), Self>
+    where
+        E: Into<KeyboardError>,
+    {
+        match err {
+            Err(err) => Err(KbdInitError {
+                cmd,
+                err: err.into(),
+            }),
+            Ok(_) => Ok(()),
         }
     }
 }
 
-/// Prints the scancode in set 1. Assumes US keyboard layout.
-// Reference: https://wiki.osdev.org/PS/2_Keyboard#Scan_Code_Set_1
+impl Display for KbdInitError {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        write!(f, "{} -  {:?}", self.cmd, self.err)
+    }
+}
+
+/// Adds the last response from the keyboard to the keyboard buffer.
+/// # Safety
+/// Reads from port 0x60 for it's response.
 #[unsafe(no_mangle)]
-extern "C" fn key_pressed_handler() {
-    let scancode = unsafe { ports::readb(Port::PS2Data) };
-    let key = unsafe { mem::transmute_copy::<u8, Key>(&scancode) };
+unsafe fn kbd_handler() {
+    /// The last value read from port 0x60.
+    static PREV_RESPONSE: AtomicU8 = AtomicU8::new(0);
 
-    // My keyboard’s enter key is permanently stuck down…
-    if cfg!(feature = "disable_enter") && key == Key::Enter {
+    if !startup::init() {
         return;
     }
 
-    match char::try_from(key) {
-        Ok(c) => print!("{c}"),
-        Err(_) => match scancode {
-            0x2A | 0x36 => KEYBOARD.shift.store(true, Ordering::Relaxed), // shift pressed
-            0xAA | 0xB6 => KEYBOARD.shift.store(false, Ordering::Relaxed), // shift released
-            0x3A => KEYBOARD.flip_caps(),                                 // caps pressed
-            0x1D => KEYBOARD.ctrl.store(true, Ordering::Relaxed),         // left ctrl released
-            0x9D => KEYBOARD.ctrl.store(false, Ordering::Relaxed),        // left ctrl released
-            0x38 => KEYBOARD.alt.store(true, Ordering::Relaxed),          // left alt pressed
-            0xB8 => KEYBOARD.shift.store(false, Ordering::Relaxed),       // left alt released
-            0x48 => vga::shift_cursor(CursorShift::Up),                   // arrow keys up
-            0x4B => vga::shift_cursor(CursorShift::Left),                 // arrow keys left
-            0x4D => vga::shift_cursor(CursorShift::Right),                // arrow keys right
-            0x50 => vga::shift_cursor(CursorShift::Down),                 // arrow keys down
-            0x0E => vga::delete_prev_char(),                              // backspace pressed
-            _ => (),
-        },
+    // Prevent another IRQ handler from adding to the buffer at the same time.
+    super::cli();
+
+    let ptr = KBD_WPTR.load(Ordering::Relaxed) as usize;
+    let scancode = unsafe { ports::readb(Port::PS2Data) };
+
+    // Save the scancode to the buffer
+    KBD_WPTR.fetch_add(1, Ordering::Relaxed);
+    KBD_BUF[ptr].store(scancode, Ordering::Relaxed);
+    PREV_RESPONSE.store(scancode, Ordering::Relaxed);
+
+    super::sti();
+}
+
+/// Polls the keyboard buffer for any new keys pressed.
+#[allow(clippy::deref_addrof)]
+pub fn poll_keyboard() {
+    /// The current state of the keyboard
+    static mut KBD: Keyboard<Us104Key, ScancodeSet2> =
+        Keyboard::new(ScancodeSet2::new(), Us104Key, HandleControl::Ignore);
+
+    // Left and right shift scancodes in set 2.
+    static LSHIFT_SCANCODE: u8 = 0x12;
+    static RSHIFT_SCANCODE: u8 = 0x59;
+
+    let read_ptr = KBD_RPTR.load(Ordering::Relaxed);
+    let write_ptr = KBD_WPTR.load(Ordering::Relaxed);
+
+    // Safety: This is the only time keyboard is mutated
+    let kbd = unsafe { &mut *&raw mut KBD };
+
+    // Return if we've reached the end of the buffer
+    if read_ptr >= write_ptr {
+        KBD_RPTR.store(write_ptr, Ordering::Relaxed);
+        return kbd.clear();
+    }
+
+    let scancode = KBD_BUF[read_ptr as usize].load(Ordering::Relaxed);
+    KBD_RPTR.fetch_add(1, Ordering::Relaxed);
+
+    // Handle shift
+    if scancode == LSHIFT_SCANCODE {
+        // Flip lshift bit
+        SHIFT.fetch_xor(1 << 0, Ordering::Relaxed);
+    } else if scancode == RSHIFT_SCANCODE {
+        // Flip rshift bit
+        SHIFT.fetch_xor(1 << 1, Ordering::Relaxed);
+    }
+
+    // If a key was pressed
+    if let Ok(event) = kbd.add_byte(scancode)
+        && let Some(event) = event
+        && event.state == KeyState::Down
+        && let Some(key) = kbd.process_keyevent(event)
+    {
+        match key {
+            DecodedKey::RawKey(key) => handle_special_key(key),
+            DecodedKey::Unicode(key) => print_key(key, kbd.get_modifiers()),
+        }
+    }
+}
+
+/// Handles when a "special" non-printable key is pressed.
+fn handle_special_key(key: KeyCode) {
+    match key {
+        KeyCode::Return | KeyCode::NumpadEnter => print!("\n"),
+        KeyCode::ArrowLeft => vga::shift_cursor(CursorShift::Left),
+        KeyCode::ArrowRight => vga::shift_cursor(CursorShift::Right),
+        KeyCode::ArrowUp => vga::shift_cursor(CursorShift::Up),
+        KeyCode::ArrowDown => vga::shift_cursor(CursorShift::Down),
+        KeyCode::Delete | KeyCode::Backspace => vga::delete_prev_char(),
+        _ => (),
+    }
+}
+
+/// Prints `key`.
+fn print_key(mut key: char, kbd: &Modifiers) {
+    /// Mapping of how to translate keys when shift is held.
+    static SHIFT_KEYS: [(char, char); 21] = [
+        ('1', '!'),
+        ('2', '@'),
+        ('3', '#'),
+        ('4', '$'),
+        ('5', '%'),
+        ('6', '^'),
+        ('7', '&'),
+        ('8', '*'),
+        ('9', '('),
+        ('0', ')'),
+        ('-', '_'),
+        ('=', '+'),
+        ('[', '{'),
+        (']', '}'),
+        ('\\', '|'),
+        (';', ':'),
+        ('\'', '"'),
+        (',', '<'),
+        ('.', '>'),
+        ('/', '?'),
+        ('`', '~'),
+    ];
+
+    // Backspace is sometimes interpreted as char 8, delete as 7F, tab as 9 and escape as 1B
+    if key == '\u{8}' || key == '\u{7F}' {
+        return vga::delete_prev_char();
+    } else if key == '\u{9}' || key == '\u{1B}' {
+        return;
+    }
+    
+    // Convert the key to it's non-shift form, to counter pc-keyboard's broken shift translation
+    let shifted = if let Some(shift) = SHIFT_KEYS.iter().find(|s| s.0 == key || s.1 == key) {
+        key = shift.0;
+        Some(shift.1)
+    } else {
+        key.make_ascii_lowercase();
+        None
     };
 
-    vga::update_vga_cursor();
+    // If shift is held
+    let shift = SHIFT.load(Ordering::Relaxed) != 0;
+
+    // Print the key in either shift, caps or regular form
+    if let Some(shifted) = shifted
+        && shift
+    {
+        print!("{shifted}")
+    } else if kbd.capslock ^ shift {
+        print!("{}", key.to_ascii_uppercase())
+    } else {
+        print!("{key}")
+    }
 }