+++ src/main.rs
@@ -1,8 +1,8 @@
 #![no_std]
 #![no_main]
-#![allow(internal_features, clippy::unusual_byte_groupings)]
+#![allow(clippy::unusual_byte_groupings)]
 #![forbid(static_mut_refs)]
-#![feature(core_intrinsics, abi_x86_interrupt)]
+#![feature(abi_x86_interrupt)]
 
 /// Allows writing to the VGA text buffer
 #[macro_use]
@@ -17,6 +17,9 @@ mod ports;
 /// Allows playing sounds through the PC speaker
 mod speaker;
 
+/// Handles post-boot startup tasks.
+mod startup;
+
 /// Handles the PIT.
 mod time;
 
@@ -29,26 +32,18 @@ run it using either `run-nosound` `run-pipewire` or `run-pulseaudio`"
 
 /// The kernel entry point.
 /// # Safety
-/// This function assumes it's being run immediately 
-/// after taking control from the bootloader.
+/// Please don't run the kernel twice.
 #[unsafe(export_name = "_start")]
 pub unsafe extern "C" fn kmain() -> ! {
-    vga::init();
-    interrupts::init();
-    time::set_timer_interval();
+    startup::run("Connected VGA", vga::init);
+    startup::run("Loaded IDT", interrupts::load_idt);
+    startup::run("Initialised PIC", interrupts::init_pic);
+    startup::run("Set PIT frequency", time::set_timer_interval);
+    startup::run("Initialised keyboard", interrupts::init_kbd);
+
+    startup::SYS_INIT.store(true, core::sync::atomic::Ordering::Relaxed);
     vga::print_color("All startup tasks completed! \u{1}\n\n", vga::Color::Green);
     vga::update_vga_cursor();
     speaker::play_chime();
-
-    // Test various errors
-    // unsafe { core::arch::asm!("int3") }         // breakpoint
-    // unsafe { core::arch::asm!("ud2") }          // invalid op
-    // panic!("Something bad happened!");          // panic
-    // unsafe { core::arch::asm!("int 42") }       // gpf
-    // unsafe { *core::ptr::dangling_mut() = 42 }; // page fault
-    // unsafe { core::arch::asm!("int 8") }        // double fault
-
-    loop {
-        unsafe { core::arch::asm!("hlt") }
-    }
+    interrupts::kbd_poll_loop()
 }